Refactor CLI main.ts to use @std/cli/parse-args, move task running logic to separate file, and add help command
